introduced TableKeyColumnManager the ConnectionManager no longer manages primary key columns or column values
